requirements.txt
colorlog==6.9.0
homeassistant==2025.2.0
pip>=21.3.1
ruff==0.11.2
pytest-homeassistant-custom-component==0.13.210